This commit stops serving HTTP traffic with axum
We don't have a good example that shows how to use WebSockets with WASM
without routing. This branch replaces the serve command with a call to
`listen_for_websockets_on`. Because BonsaiDb is no longer serving the
WASM app, it must be hosted using another server like `miniserve`. Since
miniserve defaults to 8080, the ports have been updated to ensure there
are no conflicts.